feat: modified appscenario errors

+++ apptests/appscenarios/traefik_test.go
@@ -22,7 +22,6 @@ import (
 	fluxhelmv2beta2 "github.com/fluxcd/helm-controller/api/v2beta2"
 	apimeta "github.com/fluxcd/pkg/apis/meta"
 	traefikv1a1 "github.com/traefik/traefik/v2/pkg/provider/kubernetes/crd/traefikio/v1alpha1"
-	errors "k8s.io/apimachinery/pkg/api/errors"
 )
 
 var _ = Describe("Traefik Tests", Label("traefik"), func() {
@@ -197,14 +196,18 @@ var _ = Describe("Traefik Tests", Label("traefik"), func() {
 				Expect(err).NotTo(HaveOccurred())
 
 				dashboardIngress := &networkingv1.Ingress{}
-				cl.Get(ctx, types.NamespacedName{
+				err = cl.Get(ctx, types.NamespacedName{
 					Name:      "traefik-dashboard",
 					Namespace: kommanderNamespace,
 				}, dashboardIngress)
-				Expect(err).NotTo(HaveOccurred())
-				Expect(cl.Delete(ctx, dashboardIngress)).To(Or(
-					Succeed(), MatchError(errors.IsNotFound, "IsNotFound"),
-				))
+				Expect(ctrlClient.IgnoreNotFound(err)).NotTo(HaveOccurred())
+
+				if err == nil {
+					Expect(cl.Delete(ctx, dashboardIngress)).To(Or(
+						Succeed(),
+						MatchError(Satisfy(apierrors.IsNotFound)),
+					))
+				}
 			})
 
 			By("triggering a HelmRelease reconciliation", func() {
